Refactor kube2sky.go to make it usable from other packages.
Most logic moved to lib/kube2sky_lib.go. The flag parsing logic
remains in kube2sky.go. Some structs and types were updated to
make this feasible.
